requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
pyyaml==6.0.2
scikit-learn==1.5.2
numpy==2.1.3
httpx==0.27.2
